Version 1.0 code
Add all scripts, plists and nibs for fully functional OMC applet: Delta.app
Delta.app is a wrapper around directory comparison scripts intended to generate reports from similar but not identical directories. It can be used in development for comparing different builds of the same app to see what changes were introduced: size bloat, missing files, added files. It is worth noting explicitly that for that purpose you need to be aware ot unstable  xib -> nib compiler which produces different results for the same input files. To combat the noise a binary size delta threshold can be used to filter out small insignificant size differences.

The binaries for OMCApplet, embedded Abracode.framework and replay tool are added to .gitignore so they will not take space in the repo and unnecessarily pollute the code. These binaries are coming from appropriate builds of OMC and replay and will be added in distribution archive.